Renaming system-layer into amp-story-system-layer.
