+++ extensions/amp-story/0.1/amp-story-system-layer.js
